// src/app/interfaces/interfaces.ts
export interface IAuthData {
  login: string;
  password: string;
}

export interface IUser {
  first_name: string;
  last_name: string;
  pat_name: string;
  email: string;
  birthday: string;
  phone: string;
  user_id: number;
  summ: string;
  discount: string;
  bonus: string;
  messageForUser?: string;
}